code_locations/etl_pipeline/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[project]
name = "etl_pipeline"
version = "0.1.0"
dependencies = [
    "alembic==1.18.3",
    "annotated-types==0.7.0",
    "antlr4-python3-runtime==4.13.2",
    "certifi==2026.1.4",
    "charset-normalizer==3.4.4",
    "click==8.3.1",
    "coloredlogs==14.0",
    "dagster==1.12.12",
    "dagster-duckdb==0.28.12",
    "dagster-duckdb-pandas==0.28.12",
    "dagster-pipes==1.12.12",
    "dagster-postgres==0.28.12",
    "dagster_shared==1.12.12",
    "docstring_parser==0.17.0",
    "duckdb==1.4.4",
    "filelock==3.20.3",
    "fsspec==2026.1.0",
    "greenlet==3.3.1",
    "grpcio==1.76.0",
    "grpcio-health-checking==1.76.0",
    "humanfriendly==10.0",
    "idna==3.11",
    "Jinja2==3.1.6",
    "Mako==1.3.10",
    "markdown-it-py==4.0.0",
    "MarkupSafe==3.0.3",
    "mdurl==0.1.2",
    "numpy==2.4.1",
    "packaging==26.0",
    "pandas==2.3.3",
    "pathlib_abc==0.5.2",
    "platformdirs==4.5.1",
    "protobuf==6.33.5",
    "psycopg2-binary==2.9.11",
    "pydantic==2.12.5",
    "pydantic_core==2.41.5",
    "Pygments==2.19.2",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.2.1",
    "pytz==2025.2",
    "PyYAML==6.0.3",
    "requests==2.32.5",
    "rich==14.3.1",
    "six==1.17.0",
    "SQLAlchemy==2.0.46",
    "structlog==25.5.0",
    "tabulate==0.9.0",
    "tomli==2.4.0",
    "tomlkit==0.14.0",
    "toposort==1.10",
    "tqdm==4.67.1",
    "typing-inspection==0.4.2",
    "typing_extensions==4.15.0",
    "tzdata==2025.3",
    "universal_pathlib==0.3.8",
    "urllib3==2.6.3",
    "watchdog==6.0.0"
]
